# openapi.yaml
openapi: 3.0.0
info:
  title: Complex API Demo
  version: 1.0.0
  description: A comprehensive API demonstration for Unkey deployment showcase
  contact:
    name: Unkey Team
    url: https://unkey.dev
    email: support@unkey.dev
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: /
    description: API server
components:
  schemas:
...
    Error:
      type: object
      properties:
        data:
          type: 'null'
        success:
          type: boolean
          example: false
        message:
          type: string
          example: An error occurred
...
paths:
...
  /api/users:
    get:
      summary: Get paginated list of users
      tags:
        - Users
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
          description: Number of items per page
      responses:
        '200':
          description: Paginated list of users
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          format: uuid
                        name:
                          type: string
                        email:
                          type: string
                          format: email
                        role:
                          type: string
                        status:
                          type: string
                        createdAt:
                          type: string
                          format: date-time
                  pagination:
                    type: object
                    properties:
                      page:
                        type: integer
                      limit:
                        type: integer
                      total:
                        type: integer
                      totalPages:
                        type: integer
                      hasNext:
                        type: boolean
                      hasPrev:
                        type: boolean
                  success:
                    type: boolean
                  message:
                    type: string
        '401':
          description: Unauthorized - missing or invalid authentication
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/users/{id}:
    get:
      summary: Get user by ID
      tags:
        - Users
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: User ID
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                        format: uuid
                      name:
                        type: string
                      email:
                        type: string
                        format: email
                      role:
                        type: string
                      status:
                        type: string
                      createdAt:
                        type: string
                        format: date-time
                      profile:
                        type: object
                        properties:
                          avatar:
                            type: string
                          bio:
                            type: string
                          phone:
                            type: string
                          address:
                            type: object
                  success:
                    type: boolean
                  message:
                    type: string
        '400':
          description: Bad request - invalid user ID format
        '401':
          description: Unauthorized - missing or invalid authentication
        '404':
          description: User not found
...